chore: migrate speaker data
