+++ src/components/FeaturedSpeakers.astro
@@ -1,7 +1,7 @@
 ---
 import { WebsiteConfig } from "../config";
 import eventLogo from "@assets/logo.svg";
-import { extraDetails } from "../config";
+import { extraDetails } from "../data/featured";
 
 const indicoResponse = await fetch(
   WebsiteConfig.featuredSpeakers.indicoExportUrl
