+++ src/config.ts
@@ -10,15 +10,7 @@ import PreviewImage from "@assets/ogimage.png";
 import UbuntuNepal from "@assets/ubuntu_nepal.png";
 import CfpPromoImage from "@assets/uca25cfp_promo.jpg";
 import impact_nexus from "@assets/sponsors/impact_nexus_nepal.png";
-import velascoImg from "@assets/speakers/velasco.png";
-import yuningImg from "@assets/speakers/yuning.png";
-import mauroImg from "@assets/speakers/mauro.png";
-import frankImg from "@assets/speakers/frank.png";
-import dimpleImg from "@assets/speakers/dimple.png";
-import bikalpaImg from "@assets/speakers/bikalpa.png";
-import andreeaImg from "@assets/speakers/andreea.png";
-import aadityaImg from "@assets/speakers/aaditya.png";
-import saileshImg from "@assets/speakers/sailesh.png";
+
 
 export interface SubMenuItem {
   link: string;
@@ -252,50 +244,3 @@ export const VenueFloorMaps = [
   },
 ];
 
-export const extraDetails = [
-  {
-    name: "Andreia Velasco",
-    title: "Technical Author at Canonical Data Platform",
-    image: velascoImg.src,
-  },
-  {
-    name: "Bikalpa Dhakal",
-    title: "Software Engineer at Canonical",
-    image: bikalpaImg.src,
-  },
-  {
-    name: "Dimple Kuriakose",
-    title: "Technical Author at Canonical",
-    image: dimpleImg.src,
-  },
-  {
-    name: "Mauro Gaspari",
-    title: "Community Engineer at Canonical",
-    image: mauroImg.src,
-  },
-  {
-    name: "Andreea Munteanu",
-    title: "AI/ML Manager at Canonical",
-    image: andreeaImg.src,
-  },
-  {
-    name: "Yuning Liang",
-    title: "Founder at DeepComputing",
-    image: yuningImg.src,
-  },
-  {
-    name: "Frank Karlitschek",
-    title: "Founder and CEO of Nextcloud",
-    image: frankImg.src,
-  },
-  {
-    name: "Aaditya Singh",
-    title: "Community Lead GNOME Nepal & Ubuntu Nepal",
-    image: aadityaImg.src,
-  },
-  {
-    name: "Sailesh Singh",
-    title: "System Network Administrator",
-    image: saileshImg.src,
-  },
-];
